+++ apps/client/src/environments/patch-notes.ts
@@ -1,4 +1,18 @@
 export const patchNotes = `### Bug Fixes
 
-* **core:** fixed some numbers not showing when using PT_BR locale.
-* **desktop:** rollback of recent packet capture changes to fix inventory tracking issues.`;
+* **core:** add structuredClone polyfill for older browsers.
+* **db:** fixed some content not being accessible due to italic characters in their name.
+* **desktop:** fixed omicron tokens listed in island inventory.
+* **food-picker:** now sorting results by picked stat(s) value.
+* **lists:** possible fix for final items sometimes not updating.
+* **step-by-step:** added masterbook icon next to input when required.
+* **step-by-step:** fixed arrows not accurate at all on ultrawide displays.
+* **teams:** fixed discord webhook often sending wrong values in notification.
+* **workshop:** improvements on list deletion workshop update.
+
+
+### Features
+
+* **db:** import of ~500 allagan reports.
+* **gearset:** new button to reset materia and item progression.
+* **gearset:** you can now switch from a class to its job on a gearset.`;