+++ apps/client/src/main.ts
@@ -10,6 +10,7 @@ import { enableDebugTools } from '@angular/platform-browser';
 if (environment.production) {
   enableProdMode();
 }
+enableProdMode();
 
 document.addEventListener('DOMContentLoaded', () => {
   platformBrowserDynamic().bootstrapModule(AppModule, { ngZoneEventCoalescing: true, ngZoneRunCoalescing: true }).then((module) => {